+++ CHANGELOG.rst
@@ -10,7 +10,7 @@ file.
 Added
 =====
 
-- Added endpoint ``GET v2/stored_flows`` for listing flows from flows collection. Query parameter filter for state is supported.
+- Added endpoint ``GET v2/stored_flows`` for listing flows from flows collection. Query parameter filter for `state` and `dpid` is supported. A list of flows per switch is returned.
 
 Fixed
 =====